// src/coins/tron/adapter/dto/tron-address-generate-output.dto.ts
/**
 * Adapter response payload for TRON address generation.
 */
export interface TronAddressGenerateAdapterOutput {
  address: string;
  keys: {
    public: string;
    private: string;
  };
  derivation: {
    path: string;
    purpose: number;
    coin: number;
    account: number;
    change: number;
    index: number;
  };
}